type_body_length対応の途中分を整理
