+++ YoutubeFeeder/Infrastructure/YouTube/YouTubeSearchService.swift
@@ -3,14 +3,14 @@ import Foundation
 struct YouTubeSearchService {
     nonisolated static let videoDetailsPartParameter = "snippet,contentDetails,statistics,liveStreamingDetails"
 
-    let dataLoader: @Sendable (URLRequest) async throws -> (Data, URLResponse)
+    private let transport: YouTubeSearchServiceTransport
 
     init(
         dataLoader: @escaping @Sendable (URLRequest) async throws -> (Data, URLResponse) = { request in
             try await URLSession.shared.data(for: request)
         }
     ) {
-        self.dataLoader = dataLoader
+        self.transport = YouTubeSearchServiceTransport(dataLoader: dataLoader)
     }
 
     var isConfigured: Bool {
@@ -47,15 +47,18 @@ struct YouTubeSearchService {
                 stage: &stage
             )
             stage = "video_details"
-            let detailedVideos = try await fetchVideoDetails(videoIDs: selection.videoIDs, apiKey: apiKey)
+            let detailedVideos = try await transport.fetchVideoDetails(videoIDs: selection.videoIDs, apiKey: apiKey)
             stage = "merge_details"
-            let videos = Self.mergeDetailedVideos(detailedVideos, preferredOrder: selection.videoIDs)
+            let videos = YouTubeSearchServiceProcessing.mergeDetailedVideos(detailedVideos, preferredOrder: selection.videoIDs)
             let response = YouTubeSearchResponse(videos: videos, totalCount: videos.count, fetchedAt: .now)
-            logger.info("request_complete", metadata: selection.completionMetadata(
-                keywordPreview: keywordPreview,
-                videoCount: videos.count,
-                startedAt: startedAt
-            ))
+            logger.info(
+                "request_complete",
+                metadata: selection.completionMetadata(
+                    keywordPreview: keywordPreview,
+                    videoCount: videos.count,
+                    startedAt: startedAt
+                )
+            )
             return response
         } catch {
             handleSearchFailure(error, keywordPreview: keywordPreview, stage: stage, startedAt: startedAt, logger: logger)
@@ -80,7 +83,7 @@ struct YouTubeSearchService {
         do {
             if AppLaunchMode.current.usesMockData {
                 stage = "mock_response"
-                let response = mockChannelVideosPageResponse(channelID: channelID, limit: limit)
+                let response = YouTubeSearchServiceProcessing.mockChannelVideosPageResponse(channelID: channelID, limit: limit)
                 logger.info(
                     "channel_request_complete",
                     metadata: [
@@ -95,7 +98,7 @@ struct YouTubeSearchService {
 
             let apiKey = try resolveAPIKey(keywordPreview: channelID, logger: logger)
             stage = "channel_playlist"
-            let page = try await fetchChannelVideosPage(
+            let page = try await transport.fetchChannelVideosPage(
                 channelID: channelID,
                 pageToken: pageToken,
                 limit: min(limit, 50),
@@ -142,7 +145,7 @@ struct YouTubeSearchService {
         params.stage = "mock_delay"
         try? await Task.sleep(nanoseconds: 1_000_000_000)
         params.stage = "mock_response"
-        let response = mockSearchResponse(keyword: params.keyword, limit: params.limit)
+        let response = YouTubeSearchServiceProcessing.mockSearchResponse(keyword: params.keyword, limit: params.limit)
         params.logger.info(
             "request_complete",
             metadata: [
@@ -174,21 +177,21 @@ struct YouTubeSearchService {
         stage: inout String
     ) async throws -> CandidateSelection {
         stage = "candidate_medium"
-        let mediumCandidates = try await searchCandidates(
+        let mediumCandidates = try await transport.searchCandidates(
             keyword: keyword,
             duration: "medium",
             apiKey: apiKey,
             maxResults: 50
         )
         stage = "candidate_long"
-        let longCandidates = try await searchCandidates(
+        let longCandidates = try await transport.searchCandidates(
             keyword: keyword,
             duration: "long",
             apiKey: apiKey,
             maxResults: 50
         )
         stage = "merge_candidates"
-        let mergedCandidates = Self.mergeCandidates(mediumCandidates + longCandidates)
+        let mergedCandidates = YouTubeSearchServiceProcessing.mergeCandidates(mediumCandidates + longCandidates)
         return CandidateSelection(
             mediumCandidateCount: mediumCandidates.count,
             longCandidateCount: longCandidates.count,
@@ -237,275 +240,6 @@ struct YouTubeSearchService {
 
         return nil
     }
-
-    private func searchCandidates(
-        keyword: String,
-        duration: String,
-        apiKey: String,
-        maxResults: Int
-    ) async throws -> [SearchCandidate] {
-        let logger = AppConsoleLogger.youtubeSearch
-        let startedAt = Date()
-        var components = URLComponents(string: "https://www.googleapis.com/youtube/v3/search")
-        components?.queryItems = [
-            URLQueryItem(name: "part", value: "snippet"),
-            URLQueryItem(name: "q", value: keyword),
-            URLQueryItem(name: "type", value: "video"),
-            URLQueryItem(name: "order", value: "date"),
-            URLQueryItem(name: "videoDuration", value: duration),
-            URLQueryItem(name: "videoEmbeddable", value: "true"),
-            URLQueryItem(name: "maxResults", value: String(maxResults))
-        ]
-
-        guard let url = components?.url else {
-            throw YouTubeSearchError.invalidResponse
-        }
-
-        var request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalCacheData, timeoutInterval: 30)
-        request.setValue(apiKey, forHTTPHeaderField: "X-Goog-Api-Key")
-
-        logger.info(
-            "candidate_request_start",
-            metadata: ["duration": duration, "max_results": String(maxResults), "keyword": AppConsoleLogger.sanitizedKeyword(keyword)]
-        )
-
-        let data = try await loadData(
-            for: request,
-            endpoint: "search",
-            metadata: ["duration": duration, "max_results": String(maxResults)]
-        )
-        let response = try decodeResponse(SearchListResponse.self, from: data, endpoint: "search", metadata: ["duration": duration])
-        let candidates: [SearchCandidate] = response.items.compactMap { item in
-            guard let videoID = item.id.videoID else { return nil }
-            return SearchCandidate(id: videoID, publishedAt: item.snippet.publishedAt)
-        }
-        logger.info(
-            "candidate_request_complete",
-            metadata: [
-                "duration": duration,
-                "items": String(candidates.count),
-                "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: startedAt)
-            ]
-        )
-        return candidates
-    }
-
-    private func fetchChannelVideosPage(
-        channelID: String,
-        pageToken: String?,
-        limit: Int,
-        apiKey: String,
-    ) async throws -> YouTubeChannelVideosPage {
-        let uploadsPlaylistID = try await fetchChannelUploadsPlaylistID(channelID: channelID, apiKey: apiKey)
-        let playlistItems = try await fetchPlaylistItems(
-            playlistID: uploadsPlaylistID,
-            pageToken: pageToken,
-            apiKey: apiKey,
-            maxResults: limit
-        )
-        let videoIDs = playlistItems.items.compactMap(\.contentDetails?.videoID)
-
-        let detailedVideos = try await fetchVideoDetails(videoIDs: videoIDs, apiKey: apiKey)
-        let videos = Self.mergeDetailedVideos(detailedVideos, preferredOrder: videoIDs)
-            .filter {
-                !ShortVideoMaskPolicy.shouldMask(
-                    durationSeconds: $0.durationSeconds,
-                    videoURL: $0.videoURL,
-                    title: $0.title
-                )
-            }
-        return YouTubeChannelVideosPage(
-            videos: videos,
-            totalCount: playlistItems.pageInfo?.totalResults ?? videos.count,
-            fetchedAt: .now,
-            nextPageToken: playlistItems.nextPageToken
-        )
-    }
-
-    private func fetchChannelUploadsPlaylistID(
-        channelID: String,
-        apiKey: String
-    ) async throws -> String {
-        let logger = AppConsoleLogger.youtubeSearch
-        let startedAt = Date()
-        var components = URLComponents(string: "https://www.googleapis.com/youtube/v3/channels")
-        components?.queryItems = [
-            URLQueryItem(name: "part", value: "contentDetails"),
-            URLQueryItem(name: "id", value: channelID)
-        ]
-
-        guard let url = components?.url else {
-            throw YouTubeSearchError.invalidResponse
-        }
-
-        var request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalCacheData, timeoutInterval: 30)
-        request.setValue(apiKey, forHTTPHeaderField: "X-Goog-Api-Key")
-
-        logger.info("channel_uploads_request_start", metadata: ["channelID": channelID])
-
-        let data = try await loadData(
-            for: request,
-            endpoint: "channels",
-            metadata: ["channelID": channelID]
-        )
-        let response = try decodeResponse(
-            ChannelsListResponse.self,
-            from: data,
-            endpoint: "channels",
-            metadata: ["channelID": channelID]
-        )
-        guard let uploadsPlaylistID = response.items.first?.contentDetails.relatedPlaylists.uploads,
-              !uploadsPlaylistID.isEmpty
-        else {
-            logger.error(
-                "channel_uploads_missing",
-                message: "チャンネルの uploads プレイリストを取得できませんでした。",
-                metadata: [
-                    "channelID": channelID,
-                    "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: startedAt)
-                ]
-            )
-            throw YouTubeSearchError.invalidResponse
-        }
-
-        logger.info(
-            "channel_uploads_request_complete",
-            metadata: [
-                "channelID": channelID,
-                "uploads_playlist_id": uploadsPlaylistID,
-                "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: startedAt)
-            ]
-        )
-        return uploadsPlaylistID
-    }
-
-    private func fetchPlaylistItems(
-        playlistID: String,
-        pageToken: String?,
-        apiKey: String,
-        maxResults: Int
-    ) async throws -> PlaylistItemsListResponse {
-        let logger = AppConsoleLogger.youtubeSearch
-        let startedAt = Date()
-        var components = URLComponents(string: "https://www.googleapis.com/youtube/v3/playlistItems")
-        var queryItems = [
-            URLQueryItem(name: "part", value: "contentDetails"),
-            URLQueryItem(name: "playlistId", value: playlistID),
-            URLQueryItem(name: "maxResults", value: String(max(1, min(maxResults, 50))))
-        ]
-        if let pageToken, !pageToken.isEmpty {
-            queryItems.append(URLQueryItem(name: "pageToken", value: pageToken))
-        }
-        components?.queryItems = queryItems
-
-        guard let url = components?.url else {
-            throw YouTubeSearchError.invalidResponse
-        }
-
-        var request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalCacheData, timeoutInterval: 30)
-        request.setValue(apiKey, forHTTPHeaderField: "X-Goog-Api-Key")
-
-        logger.info(
-            "playlist_items_request_start",
-            metadata: [
-                "playlist_id": playlistID,
-                "page_token": pageToken ?? "",
-                "max_results": String(max(1, min(maxResults, 50)))
-            ]
-        )
-
-        let data = try await loadData(
-            for: request,
-            endpoint: "playlistItems",
-            metadata: [
-                "playlist_id": playlistID,
-                "page_token": pageToken ?? "",
-                "max_results": String(max(1, min(maxResults, 50)))
-            ]
-        )
-        let response = try decodeResponse(
-            PlaylistItemsListResponse.self,
-            from: data,
-            endpoint: "playlistItems",
-            metadata: ["playlist_id": playlistID, "page_token": pageToken ?? ""]
-        )
-        logger.info(
-            "playlist_items_request_complete",
-            metadata: [
-                "playlist_id": playlistID,
-                "page_token": pageToken ?? "",
-                "items": String(response.items.count),
-                "next_page_token": response.nextPageToken ?? "",
-                "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: startedAt)
-            ]
-        )
-        return response
-    }
-
-    func fetchVideoDetails(videoIDs: [String], apiKey: String) async throws -> [YouTubeSearchVideo] {
-        guard !videoIDs.isEmpty else { return [] }
-
-        let logger = AppConsoleLogger.youtubeSearch
-        let startedAt = Date()
-        var mergedVideos: [YouTubeSearchVideo] = []
-        let batches = chunkVideoIDs(videoIDs, size: 50)
-        logger.info(
-            "video_details_start",
-            metadata: ["video_ids": String(videoIDs.count), "batches": String(batches.count)]
-        )
-        for (index, batch) in batches.enumerated() {
-            var components = URLComponents(string: "https://www.googleapis.com/youtube/v3/videos")
-            components?.queryItems = [
-                URLQueryItem(name: "part", value: Self.videoDetailsPartParameter),
-                URLQueryItem(name: "id", value: batch.joined(separator: ",")),
-                URLQueryItem(name: "maxResults", value: String(batch.count))
-            ]
-
-            guard let url = components?.url else {
-                throw YouTubeSearchError.invalidResponse
-            }
-
-            var request = URLRequest(url: url, cachePolicy: .reloadIgnoringLocalCacheData, timeoutInterval: 30)
-            request.setValue(apiKey, forHTTPHeaderField: "X-Goog-Api-Key")
-
-            let data = try await loadData(
-                for: request,
-                endpoint: "videos",
-                metadata: ["batch": "\(index + 1)/\(batches.count)", "ids": String(batch.count)]
-            )
-            let response = try decodeResponse(
-                VideoListResponse.self,
-                from: data,
-                endpoint: "videos",
-                metadata: ["batch": "\(index + 1)/\(batches.count)"]
-            )
-            let videos = Self.filterPlayableVideos(response.items)
-            mergedVideos.append(contentsOf: videos)
-            logger.info(
-                "video_details_batch_complete",
-                metadata: ["batch": "\(index + 1)/\(batches.count)", "videos": String(videos.count)]
-            )
-        }
-
-        logger.info(
-            "video_details_complete",
-            metadata: ["videos": String(mergedVideos.count), "elapsed_ms": AppConsoleLogger.elapsedMilliseconds(since: startedAt)]
-        )
-        return mergedVideos
-    }
-
-    private func chunkVideoIDs(_ videoIDs: [String], size: Int) -> [[String]] {
-        guard size > 0 else { return [videoIDs] }
-        var result: [[String]] = []
-        var index = 0
-        while index < videoIDs.count {
-            let nextIndex = min(index + size, videoIDs.count)
-            result.append(Array(videoIDs[index ..< nextIndex]))
-            index = nextIndex
-        }
-        return result
-    }
-
 }
 
 private struct CandidateSelection {
